remove unused code

+++ Assets/distort_map.shader
@@ -46,9 +46,8 @@
             fixed4 frag(v2f i) : SV_Target
             {
                 float4 data = tex2D(_DistortTex, i.uv);
-                fixed4 col = tex2D(_MainTex, float2(data.x,data.y));
+                float4 col = tex2D(_MainTex, float2(data.x,data.y));
                 return col * col.a + tex2D(_CamTex, i.uv) * (1 - col.a);
-                //return col;
             }
             ENDCG
         }
